isolate buffer based algorithms into buffer_algorithms.hpp, remove std::plus by plus (idem with multiplies), merge descriptors into sycl_algorithm_descriptor
