+++ include/sycl/algorithm/exclusive_scan.hpp
@@ -31,7 +31,7 @@
 
 #include <sycl/helpers/sycl_buffers.hpp>
 #include <sycl/helpers/sycl_namegen.hpp>
-#include <sycl/algorithm/inclusive_scan.hpp>
+#include <sycl/algorithm/buffer_algorithms.hpp>
 
 namespace sycl {
 namespace impl {
