+++ include/sycl/algorithm/inclusive_scan.hpp
@@ -30,7 +30,7 @@
 #define __SYCL_IMPL_ALGORITHM_INCLUSIVE_SCAN__
 
 #include <sycl/helpers/sycl_buffers.hpp>
-#include <sycl/algorithm/reduce.hpp>
+#include <sycl/algorithm/buffer_algorithms.hpp>
 
 namespace sycl {
 namespace impl {
@@ -90,7 +90,7 @@ OutputIterator inclusive_scan(ExecutionPolicy &sep, InputIterator b,
       auto aO =
           outBuf->template get_access<cl::sycl::access::mode::read_write>(h);
       h.parallel_for<typename ExecutionPolicy::kernelName>(
-          r, [aI, aO, bop, vectorSize, i](cl::sycl::item<1> id) {
+          r, [aI, aO, bop, vectorSize, i](cl::sycl::nd_item<1> id) {
             size_t td = 1 << (i - 1);
             size_t m_id = id.get_global(0);
 
@@ -110,231 +110,6 @@ OutputIterator inclusive_scan(ExecutionPolicy &sep, InputIterator b,
   return o + vectorSize;
 }
 #else
-
-
-struct mapscan_descriptor {
-  size_t size, size_per_work_group, size_per_work_item,
-    nb_work_group, nb_work_item;
-  mapscan_descriptor(size_t size_,
-                     size_t size_per_work_group_,
-                     size_t size_per_work_item_,
-                     size_t nb_work_group_,
-                     size_t nb_work_item_):
-    size(size_),
-    size_per_work_group(size_per_work_group_),
-    size_per_work_item(size_per_work_item_),
-    nb_work_group(nb_work_group_),
-    nb_work_item(nb_work_item_) {}
-};
-
-mapscan_descriptor compute_mapscan_descriptor(cl::sycl::device device,
-                                              size_t size,
-                                              size_t sizeofB) {
-  //std::cout << "size=\t" << size << std::endl;
-  using std::min;
-  using std::max;
-  if (size == 0)
-    return mapscan_descriptor(0, 0, 0, 0, 0);
-  size_t local_mem_size =
-    device.get_info<cl::sycl::info::device::local_mem_size>();
-  //std::cout << "local_mem_size=\t" << local_mem_size << std::endl;
-  size_t size_per_work_group = min(size, local_mem_size / sizeofB);
-  //std::cout << "size_per_work_group=\t" << size_per_work_group << std::endl;
-  if (size_per_work_group <= 0)
-    return mapscan_descriptor(size, 0, 0, 0, 0);
-
-  size_t nb_work_group = up_rounded_division(size, size_per_work_group);
-  //std::cout << "nb_work_group=\t" << nb_work_group << std::endl;
-
-  size_t max_work_item =
-    device.get_info<cl::sycl::info::device::max_work_group_size>();
-  //std::cout << "max_work_item=\t" << max_work_item << std::endl;
-  size_t nb_work_item = min(max_work_item, size_per_work_group);
-  //std::cout << "nb_work_item=\t" << nb_work_item << std::endl;
-  size_t size_per_work_item =
-    up_rounded_division(size_per_work_group, nb_work_item);
-  //std::cout << "size_per_work_item=\t" << size_per_work_item << std::endl;
-  return mapscan_descriptor(size, size_per_work_group,
-                            size_per_work_item, nb_work_group, nb_work_item);
-}
-
-
-template <class ExecutionPolicy, class A, class B, class Reduce, class Map>
-void buffer_mapscan(ExecutionPolicy &snp,
-                    cl::sycl::queue q,
-                    cl::sycl::buffer<A, 1> input_buffer,
-                    cl::sycl::buffer<B, 1> output_buffer,
-                    B init,
-                    mapscan_descriptor d,
-                    Map map,
-                    Reduce red) {
-    //map is not applied on init
-
-  using std::min;
-  using std::max;
-
-  size_t size = d.size;
-  size_t size_per_work_group = d.size_per_work_group;
-  size_t size_per_work_item = d.size_per_work_item;
-  size_t nb_work_group = d.nb_work_group;
-  size_t nb_work_item = d.nb_work_item;
-
-  //WARNING: nb_work_group is not bounded by max_compute_units
-  cl::sycl::buffer<B, 1> scan = { cl::sycl::range<1> { nb_work_group } };
-  cl::sycl::range<1> rng_wg {nb_work_group * nb_work_item};
-  cl::sycl::range<1> rng_wi {nb_work_item};
-
-  q.submit([&] (cl::sycl::handler &cgh) {
-    auto input =
-      input_buffer.template get_access<cl::sycl::access::mode::read>(cgh);
-    auto output =
-      output_buffer.template get_access<cl::sycl::access::mode::write>(cgh);
-
-    cl::sycl::accessor<B, 1, cl::sycl::access::mode::read_write,
-                       cl::sycl::access::target::local>
-      scratch { cl::sycl::range<1> { size_per_work_group }, cgh };
-
-    cgh.parallel_for_work_group<class wg>(rng_wg, rng_wi, [=](cl::sycl::group<1> grp) {
-      //assert(false);
-      //B scratch[size_per_work_group];
-      size_t group_id = grp.get(0);
-      assert(group_id < nb_work_group);
-      size_t group_begin = group_id * size_per_work_group;
-      size_t group_end   = min((group_id+1) * size_per_work_group, size);
-      size_t local_size = group_end - group_begin;
-      assert(group_begin < group_end); //< as we properly selected the
-                                       //  number of work_group
-
-      // Step 0:
-      // each work_item copy a piece of data
-      // map is applied during the process
-      parallel_for_work_item(grp, [&](cl::sycl::item<1> id) {
-        size_t local_id  = id.get(0) % nb_work_item;
-        // gpos: position in the global vector
-        // lpos: position in the local vector
-        for (size_t gpos = group_begin + local_id, lpos = local_id;
-            gpos < group_end;
-            gpos+=nb_work_item, lpos+=nb_work_item) {
-          scratch[lpos] = map(input[gpos]);
-        }
-      });
-
-      // Step 1:
-      // each work_item scan a piece of data
-      parallel_for_work_item(grp, [&](cl::sycl::item<1> id) {
-        size_t local_id  = id.get(0) % nb_work_item;
-        size_t local_pos = local_id * size_per_work_item;
-        size_t local_end = min((local_id+1) * size_per_work_item, local_size);
-        if (local_pos < local_end) {
-          B acc = scratch[local_pos];
-          local_pos++;
-          for (; local_pos < local_end; local_pos++) {
-            acc = red(acc, scratch[local_pos]);
-            scratch[local_pos] = acc;
-          }
-        }
-      });
-
-      // Step 2:
-      {
-        // scan on every last item
-        size_t local_pos = size_per_work_item - 1;
-        if (local_pos < local_size)
-        {
-          B acc = scratch[local_pos];
-          local_pos += size_per_work_item;
-          for (; local_pos < local_size; local_pos += size_per_work_item) {
-            acc = red(acc, scratch[local_pos]);
-            scratch[local_pos] = acc;
-          }
-        }
-      }
-
-      // Step 3:
-      // (except for group = 0) add the last element of the previous block
-      parallel_for_work_item(grp, [&](cl::sycl::item<1> id) {
-        size_t local_id  = id.get(0) % nb_work_item;
-        if (local_id > 0) {
-          size_t local_pos = local_id * size_per_work_item;
-          size_t local_end = min((local_id+1) * size_per_work_item - 1,
-                                 local_size);
-          if (local_pos < local_end) {
-            B acc = scratch[local_pos - 1];
-            for (; local_pos < local_end; local_pos++) {
-              scratch[local_pos] = red(acc, scratch[local_pos]);
-            }
-          }
-        }
-      });
-
-      // Step 4:
-      // each work_item copy a piece of data
-      parallel_for_work_item(grp, [&](cl::sycl::item<1> id) {
-        size_t local_id = id.get(0) % nb_work_item;
-        size_t gpos = group_begin + local_id;
-        // gpos: position in the global vector
-        size_t lpos = local_id;
-        // lpos: position in the local vector
-        for (size_t gpos = group_begin + local_id, lpos = local_id;
-            gpos < group_end;
-            gpos+=nb_work_item, lpos+=nb_work_item) {
-          output[gpos] = scratch[lpos];
-        }
-      });
-
-    });
-  });
-
-  // STEP II: global scan
-  {
-    auto buff  = output_buffer.template get_access
-      <cl::sycl::access::mode::read_write,
-       cl::sycl::access::target::host_buffer>();
-    auto write_scan  = scan.template get_access
-      <cl::sycl::access::mode::write, cl::sycl::access::target::host_buffer>();
-    B acc = init;
-    for (size_t global_pos = size_per_work_group - 1, local_pos = 0;
-        local_pos < nb_work_group - 1;
-        local_pos++, global_pos+=size_per_work_group) {
-      write_scan[local_pos] = acc;
-      acc = red(acc, buff[global_pos]);
-    }
-    write_scan[nb_work_group - 1] = acc;
-  }
-
-
-  // STEP III: propagate global scan on local scans
-  q.submit([&] (cl::sycl::handler &cgh) {
-    auto buff = output_buffer.template get_access
-      <cl::sycl::access::mode::read_write>(cgh);
-    auto read_scan = scan.template get_access
-      <cl::sycl::access::mode::read>(cgh);
-
-    cgh.parallel_for_work_group<class wg>(rng_wg, rng_wi, [=](cl::sycl::group<1> grp) {
-      size_t group_id = grp.get(0);
-      B acc = read_scan[group_id];
-      assert(group_id < nb_work_group);
-      size_t group_begin = group_id * size_per_work_group;
-      size_t group_end   = min((group_id+1) * size_per_work_group, size);
-      assert(group_begin < group_end); //< as we properly selected the
-                                       //  number of work_group
-
-      parallel_for_work_item(grp, [&](cl::sycl::item<1> id) {
-        size_t local_id = id.get(0) % nb_work_item;
-        // gpos: position in the global vector
-        // lpos: position in the local vector
-        for (size_t gpos = group_begin + local_id;
-             gpos < group_end;
-             gpos += nb_work_item) {
-          buff[gpos] = red(acc, buff[gpos]);
-        }
-      });
-    });
-
-  });
-
-  return;
-}
 
 template <class ExecutionPolicy, class InputIterator, class OutputIterator,
           class T, class BinaryOperation>
